Ubuntu 14.04 LTS already has zeromq packages
We no longer install zeromq via PPA as the Trusty LTS release already has
zeromq libraries. The only thing missing are the PHP bindings. Those are
still installed via pecl (but with sudo rights as otherwise write
permissions of pecl are not sufficient).

From now on the hb-showcase puppet recipes are no longer running on Ubuntu
12.04 up to 13.10 as the installation via PPA has been removed. As we want
to use Ubuntu 14.04 LTS as the base for our development boxes this should
not be a problem.