almost running application now - XSDs now have to
be maintained in honeybee and in this honeybee-project-template
(honeybee-showcase) as symlinking the schema files did not work due to
relative paths in schemaLocation attributes.

The User module is now symlinked like the Core module of honeybee. This
works for the moment and schould perhaps be replaced with a symlinking
mechanism that adds honeybee modules as vendor libs and symlinks them into
the app/modules directory.

refs #21457