copy routing_module_defaults.xml file to project
on `make link-project`. The default routing file is .gitignore'd to make
it clear, that this file should not be changed. Developers can override
the default routes they don't like in the routing.xml file of the project
or the modules.

GenOnly routes can be overridden in the app/config/routing.xml file while
module specific routes can be overwritten in the module's routing.xml
file.

Bumped Honeybee and other vendor library versions.

refs #22150